Fix version pinning when setup_requires is a list
This fixes a bug which would cause the distutils hook not to be
installed if the `setupmeta` entry in `setup_requires` had been
pinned/constrained with a version spec AND the value of `setup_requires`
was a list.

Note that if the value of `setup_requires` had been specified as a
string, (e.g., `setup_requires='setupmeta==2.6.12'`) then everything
would still work as expected.